refactor: move instructions.rs into compiler.rs
